Move geometric operations in interpolate to space
